+++ Packages/MinnaChat/Sources/MinnaChat/Log.swift
@@ -1,0 +1,12 @@
+//
+//  Log.logger.swift.swift
+//  MinnaChat
+//
+//  Created by Taylor Lineman on 8/25/26.
+//
+
+import Logging
+
+enum Log {
+    nonisolated static let logger: Logger = Logger(label: "com.tryminna.minna.minnaChat")
+}